RationalNumber.java

+++ RationalNumber.java
@@ -10,7 +10,7 @@ public class RationalNumber extends RealNumber{
   }
 
   public double getValue(){
-    return numerator / denominator;
+    return (double) numerator / denominator;
   }
 
   public int getNumerator(){
